+++ ninja-servlet-jpa-integration-test/src/main/java/conf/Module.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2013 the original author or authors.
+ * Copyright (C) 2012 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,11 +17,15 @@
 package conf;
 
 import com.google.inject.AbstractModule;
+import com.google.inject.Singleton;
 
+@Singleton
 public class Module extends AbstractModule {
+    
 
-    @Override
     protected void configure() {
+        
+        bind(StartupActions.class);    
         
     }