+++ ninja-servlet-jpa-integration-test/src/test/java/controllers/ApplicationControllerFluentLeniumTest.java
@@ -17,30 +17,33 @@
 package controllers;
 
 import static org.junit.Assert.assertTrue;
-import ninja.NinjaTest;
+import ninja.NinjaFluentLeniumTest;
 
 import org.junit.Before;
 import org.junit.Test;
 
-public class ApplicationControllerTest extends NinjaTest {
+public class ApplicationControllerFluentLeniumTest extends NinjaFluentLeniumTest {
     
     @Before
     public void setup() {
         
-        ninjaTestBrowser.makeRequest(getServerAddress() + "setup");
+        goTo(getServerAddress() + "setup");
         
     }
 
     @Test
     public void testThatHomepageWorks() {
+        
+        goTo(getServerAddress() + "/");
+        
+        System.out.println("title: " + title());
+        
+        assertTrue(title().contains("Home page"));
+        
+        click("#login");
+        
+        assertTrue(url().contains("login"));
 
-        // /redirect will send a location: redirect in the headers
-        String result = ninjaTestBrowser.makeRequest(getServerAddress() + "/");
-
-        // If the redirect has worked we must see the following text
-        // from the index screen:
-        assertTrue(result.contains("Hello to the blog example!"));
-        assertTrue(result.contains("My second post"));
 
     }