+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -16,6 +16,8 @@ Version X.X.X
  * 2013-09-12 Added https://code.google.com/p/error-prone/ to prevent bugs at compile time. (ra)
  * 2013-09-12 Now using module.setDefaultUseWrapper(false); in xml rednering 
               (see also https://github.com/FasterXML/jackson-dataformat-xml) (ra)
+ * 2013-09-13 Improved documentation (testing, controller) (ra)
+ * 2013-09-14 Added blog integration test to JPA (ra)
 
 Version 1.6.0
 =============